# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Class API",
        "version": "2"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "description": "Authenticate and login user based on personal id",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "personal_id"
                                ],
                                "properties": {
                                    "personal_id": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "404": {
                        "description": "משתמש אינו נמצא"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        },
        "/api/rahtal": {
            "get": {
                "tags": [
                    "Rahtal"
                ],
                "summary": "Get Rahtal",
                "description": "Returns the Rahtal details",
                "operationId": "rahtal",
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "404": {
                        "description": "משתמש אינו נמצא"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        },
        "/api/rahtal/{uuid}": {
            "post": {
                "tags": [
                    "Rahtal"
                ],
                "summary": "Update the rahtal",
                "description": "Update a rahtal",
                "operationId": "update rahtal",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the rahtal",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "רחטל"
                                    },
                                    "image": {
                                        "description": "The image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "רחטל עודכן בהצלחה"
                    },
                    "400": {
                        "description": "הכנס לפחות שדה אחד לעדכון"
                    },
                    "404": {
                        "description": "רחטל לא נמצא"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get authenticated user",
                "description": "Returns the authenticated user's details",
                "operationId": "user",
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "401": {
                        "description": "המשתמש לא מחובר"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        },
        "/api/users/admins": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve all admins",
                "description": "Retrieve all admins",
                "operationId": "index user",
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        },
        "/api/users/{uuid}": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "add admin",
                "description": "add admin",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "404": {
                        "description": "משתמש אינו נמצא"
                    },
                    "409": {
                        "description": "משתמש זה הינו מנהל מערכת"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "delete admin",
                "description": "delete admin role",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the admin",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "404": {
                        "description": "משתמש אינו נמצא"
                    },
                    "409": {
                        "description": "המשתמש אינו מנהל מערכת"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        },
        "/api/users/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve all users but admins",
                "description": "Retrieve all users but admins",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        },
        "/api/websites": {
            "get": {
                "tags": [
                    "Websites"
                ],
                "summary": "Retrieve all websites",
                "description": "Retrieve all websites",
                "operationId": "index websites",
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            },
            "post": {
                "tags": [
                    "Websites"
                ],
                "summary": "Upload a website",
                "description": "Upload a website",
                "operationId": "storeWebsite",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "link",
                                    "image",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "אתר 1"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://www.google.com/"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "אתר טכנולוגיה"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        },
        "/api/websites/{uuid}": {
            "post": {
                "tags": [
                    "Websites"
                ],
                "summary": "Update a website",
                "description": "Update a website",
                "operationId": "updateWebsite",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the website",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "אתר 1"
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://www.google.com/"
                                    },
                                    "image": {
                                        "description": "The image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "אתר טכנולוגיה"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "404": {
                        "description": "האתר לא נמצא"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Websites"
                ],
                "summary": "delete website",
                "description": "delete website",
                "operationId": "delete website",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of the website",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה התבצעה בהצלחה"
                    },
                    "204": {
                        "description": "לא נדרש לבצע את הפעולה"
                    },
                    "404": {
                        "description": "אתר לא נמצא"
                    },
                    "500": {
                        "description": "אירעה שגיאה"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Rahtal",
            "description": "Rahtal"
        },
        {
            "name": "Users",
            "description": "Users"
        },
        {
            "name": "Websites",
            "description": "Websites"
        }
    ]
}